edbs/pyproject.toml
[tool.poetry]
name = "edbs"
version = "0.1.0"
description = ""
authors = ["Patrick Stünkel <webminz@googlemail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
antlr4-python3-runtime = "^4.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
